requirements.txt
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
backports.weakref==1.0.post1
brotlipy==0.7.0
constantly==15.1.0
cssselect==1.1.0
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-import-export==4.1.1
django-rest-passwordreset==1.4.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.25.0
inflection==0.5.1
jsonpointer==2.1
munkres==1.1.4
oauthlib==3.2.2
openpyxl==3.0.10
patsy==0.5.3
pbr==6.0.0
pep8==1.7.1
pipenv==2024.0.1
ply==3.11
pyarrow==11.0.0
pyasn1-modules==0.2.8
PyDispatcher==2.0.5
PyJWT==2.8.0
pyls-spyder==0.4.0
PyQt5-sip==12.11.0
python-lsp-jsonrpc==1.0.0
python3-openid==3.2.0
queuelib==1.5.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stevedore==5.2.0
tablib==3.5.0
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
webencodings==0.5.1